Resolve of merge (conflict) 3c848194f28d CodeGen: Expand memory intrinsics in PreISelIntrinsicLower.. 'main' -> 'xmain-web' (#13833)
CONFLICT (content): Merge conflict in
llvm/lib/CodeGen/PreISelIntrinsicLowering.cpp

  commit 3c848194f28decca41b7362f9dd35d4939797724
  Author: Matt Arsenault <Matthew.Arsenault@amd.com>
  Date:   Wed Jun 7 09:03:17 2023 -0400

      CodeGen: Expand memory intrinsics in PreISelIntrinsicLowering

      Expand large or unknown size memory intrinsics into loops in the
default lowering pipeline if the target doesn't have the corresponding
libfunc. Previously AMDGPU had a custom pass which existed to call the
      expansion utilities.

With a default no-libcall option, we can remove the libfunc checks in
LoopIdiomRecognize for these, which never made any sense. This also
      provides a path to lifting the immarg restriction on
      llvm.memcpy.inline.

There seems to be a bug where TLI reports functions as available if
      you use -march and not -mtriple.

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Ji, Jinsong <jinsong.ji@intel.com>