[VPlan][OptRpt] Don't emit remarks for statistics which are '0'. (#13786)
We have a number of statistics which are tracked in the opt-report, e.g:

    remark #15447: --- begin vector loop memory reference summary ---
    remark #15450: unmasked unaligned unit stride loads: 1
    remark #15451: unmasked unaligned unit stride stores: 1
    remark #15456: masked unaligned unit stride loads: 0
    remark #15457: masked unaligned unit stride stores: 0
    remark #15458: masked indexed (or gather) loads: 0
    remark #15459: masked indexed (or scatter) stores: 0
    remark #15462: unmasked indexed (or gather) loads: 0
    remark #15463: unmasked indexed (or scatter) stores: 0
remark #15554: Unmasked VLS-optimized loads (each part of the group
counted separately): 0
remark #15555: Masked VLS-optimized loads (each part of the group
counted separately): 0
remark #15556: Unmasked VLS-optimized stores (each part of the group
counted separately): 0
remark #15557: Masked VLS-optimized stores (each part of the group
counted separately): 0
    remark #15497: vector compress: 0
    remark #15498: vector expand: 0
    remark #15474: --- end vector loop memory reference summary ---

As one may notice, many of these stats may not be relevant (i.e. have a
value of '0') to the loop being vectorized. As such, their presence
simply clutters the opt-report. This situation worsens as we continue to
add more stats. Moreover, it is arguably not useful to a user to know
what kinds of load/stores their loop did *not* contain.

This patch removes that behavior and now skips remark stats which would
just be zero. This results in the following (much-cleaner) report:

    remark #15447: --- begin vector loop memory reference summary ---
    remark #15450: unmasked unaligned unit stride loads: 1
    remark #15451: unmasked unaligned unit stride stores: 1
    remark #15474: --- end vector loop memory reference summary ---

The above report contains only the stats which were actually relevant to
the vectorized loop, greatly reducing the amount of noise.