[VPlan][CG] Don't emit an aligned load for the peeled memref if peeling can be skipped. (#13427)
When dynamically peeling, we may emit a run-time trip count check (if
`-vplan-calc-min-profitable-dyn-peel` or `-vplan-skip-dynamic-peel-tc`
are specified) to skip the peeling loop. In event we do so, it is not
valid to emit an aligned load for the peeled memref, as in the case that
peeling is skipped, it is possible for the loaded address to not be
aligned. If this is the case, the emitted code will segfault at
run-time.

To avoid this, before executing the plan, we perform a check to see if
the minimum profitable peel trip count is set. If so, we toggle a flag
in VPOCodeGenHIR to disable the emission of an aligned load on the
peeled memref.