[llvm-profgen] Add --perf-event to filter by perf events (#13477)
This change teaches LBR parsing to recognize the "event" field as
emitted by perf-script(1). The field is optional in order to preserve
compatibility with existing perf-script files.

When converting to perf script output the "event" field is now always
requested so that we can filter samples on it.

Two new options, are added in order to request the filtering:
`--perf-event` and `--perf-events`. The singular spelling appends a
single event name to the set of events which will be used. The plural
spelling accepts a comma-separated list.

The following examples show how this can accommodate either short
symbolic events or events with parameter lists:

- `llvm-profgen --leading-ip-only
--perf-events=mem-loads:upp,mem-stores:upp`
- `llvm-profgen --leading-ip-only
--perf-event=cpu/event=0xcd,umask=0x1,ldlat=0x3/upp
--perf-event=cpu/event=0xd0,umask=0x82/upp`

This change makes it possible to generate multiple kinds of profiles
from a single `perf record` collection. For example, to generate both
block frequency and unpredictable branch profiles:

$ perf record -c 1000003 -b -e
br_inst_retired.near_taken:upp,br_misp_retired.all_branches:upp ...
$ llvm-profgen --output hotness.prof
--perf-event=br_inst_retired.near_taken:upp ...
$ llvm-profgen --leading-ip-only --output unpredictable.prof
--perf-event=br_misp_retired.all_branches:upp ...