[NFC] Print dimension element type instead of dimension type (#13924)
With the move to opaque pointers, dimension types are just 'ptr' in many
cases
so they don't convey any information. Moreover, we print number of
elements in
the dimension as well so printing array dimension type doesn't provide
much
extra information. In the future, we may completely eliminate the
dimension type
from the ref.