[VPlan] Account for unrolling in VPLoopInfo::getInduction(). (#13874)
This patch fixes a crash caused by an interaction between
`VPLoopInfo::getInduction()` and unrolling. The previous implementation
of this method worked by matching on the latch update to find the
corresponding induction PHI, i.e. we assumed all latch updates were of
the form:

    <update> = <bin-op> <PHI> <...> or
    <update> = <bin-op> <...> <PHI>

While this is generally true, this is not the case if the loop is
unrolled. In that case, the latch update will be of the form:

    <update> = <bin-op> <prev-update> <...> or
    <update> = <bin-op> <...> <prev-update>

where `<prev-update>` is a duplicated version of the update instruction.
These updates are cascaded the same number of times as the loop has been
unrolled.

This patch splits this method into two parts:
1. `VPLoopInfo::getInductionPHI()` which handles the above matching
       logic to correctly identify the loop PHI, even if the loop has
       been unrolled.
2. `VPLoopInfo::getInductionInit()` which simply obtains the
       `VPInductionInit` corresponding to the incoming value from the
       loop PHI.

This fixes the crash and allows dynamic peeling to work properly on
unrolled loops (as the CFG merger relies on this method.)

JR: CMPLRLLVM-48619.