[VPlan] Add INTERNAL macro for hiding strings from release compilers (#13879)
Add a macro that suppresses strings that are not needed in release
compilers so that information in those strings is not leaked to the
release binary.  Such strings are also unnecessary to translate into
other languages, so marking them as INTERNAL emphasizes this.  Make
use of the macro in handling bail-out strings in the vectorizer.
(Some of the language in HIR code-generation bail-out strings was
the motivation for this patch.)

I've applied INTERNAL to debug strings in bailoutWithDebug() calls,
as well as to bail-out remark strings with verbosity High.  For both
classes of strings, these are not intended for use in release compilers.

This patch is NFC for non-release compilers, and effectively NFC for
release compilers other than causing the strings to be stripped.

Reference: CMPLRLLVM-46365