[CMPLRLLVM-47161] Add LoopStatistics for Region (#13657)
Original JR reported high compile times for a simple IVDep testcase
where DSE was taking too long. After some triaging, it was found that
DSE was calling domination utility at the region level. Because loop
statistics does not cache region-level info, we were not early exiting
and running expensive logic to check all region refs. With this change,
we are able to speedup domination checks for regions to reasonable
times.