Added x86-64 in X86TargetParser.def (#13949)
The default processor used in the Fortran front end is "x86-64". The
processor tables in X86TargetProcessor.cpp also uses "x86-64" as the
default 64-bit x86 processor. Adding an "x86-64" entry here allows the
Fortran front end to get a valid result from
`llvm::X86::VectorAbiIsaInfo::getByName("x86-64")`.

This table in X86TargetParser.def will be merged with the table in
X86TargetParser.cpp soon, and when that change happens "x86-64" will be
defined in the new combined table. Though this commit will eventually be
superseded, merging it now allows the Fortran front end to enable
cpu-dispatch vector variants before the upstream change merges the
tables. That way the front end and back end code can be tested together
sooner and more thoroughly than if we wait for the cleaner change of
merging tables.

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>

---------

Signed-off-by: William R. Dieter <william.r.dieter@intel.com>