Resolve of merge (conflict) adcdc9cc3740 [LLD][COFF] Allow overwriting directives exports with cmd-.. 'main' -> 'xmain-web' (#14024)
CONFLICT (content): Merge conflict in lld/COFF/Driver.cpp

  commit adcdc9cc3740adba3577b328fa3ba492cbccd3a5
  Author: Alexandre Ganea <alex_toresh@yahoo.fr>
  Date:   Tue Jun 13 15:12:55 2023 -0400

[LLD][COFF] Allow overwriting directives exports with cmd-line exports

MSVC link.exe allows overriding exports on the cmd-line with exports
seen in OBJ directives. The typical case is what is described in #62329.

Before this patch, trying to override an export with `/export` or `/def`
would generate a duplicate warning. This patches tries to replicate the
MSVC behavior. A second override on the cmd-line would still generate
the warning.

There's still a case which we don't cover: MSVC link.exe is able to
demangle an exported OBJ directive function, and match it with a
unmangled export function in a .def file. In the meanwhile, one can use
the mangled export in the .def to cover that case.

      This fixes #62329

      Differential revision: https://reviews.llvm.org/D149611

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: sys_ce_bb <sys_ce_bb@intel.com>