[VPlan] Enable min profitable dynamic peel tc calculation by default. (#13919)
PR #13418 added the switch `-vplan-calc-min-profitable-dyn-peel-tc` to
control whether the vectorizer would compute and emit a profitable trip
count check when dynamic peeling, with a default value of `false`.

This patch enables the behavior by default (changes the default switch
value to `true`). Additionally, we slightly modify the behavior to only
emit a peel check if the trip count is unknown.

JR: CMPLRLLVM-47569.