[VPlan] Fix crash in profitable peel threshold computation. (#14066)
When dynamically peeling, we compute a profitable peel threshold to emit
a dynamic TC check. This is done by computing the ratio of two
quantities which represent, roughly: the overhead cost of aligning the
loop without considering trip-count (i.e. the numerator), and the
expected gain from alignment for each iteration (i.e. the denominator.)

We had an assert to check that the denominator was not negative, as well
as an assert that the entire computation (`numerator/denominator`) was
not negative, however there was no check for the numerator being
negative. This turns out to represent a valid scenario: one where the
overhead "cost" of alignment is negative -- meaning that the aligned
loop is always better than the unaligned loop.

This issue was found in a SYCL benchmark (`sgemm-intel`) where enabling
dynamic peeling and vectorized peel led to the numerator being negative,
causing the `assert` to fire as described above.

This patch fixes the crash by computing a value of '0' for the
profitable trip count in this case -- indicating that no TC check will
be emitted and that peeling will proceed unconditionally.

Additionally, a typo was fixed in the computation of the numerator (a
`+` that should have been a `-`).