[NFC] Disable/modify tests which do not work with new layout (#14050)
INTEL_DEPLOY_UNIFIED_LAYOUT causes the driver to use a different
expected directory structure.  Some tests do not support this layout and
need to be disabled under these conditions.  These tests should be
updated to work with INTEL_DEPLOY_UNIFIED_LAYOUT once it is made the
default.

Also, update the unified test that was checking the wrong directories
that are set by the driver.