Use gotos cached by HIRLoopStatistics in domination utility to make it more efficient and precise (#14057)
This PR makes domination utility completely dependent on
HIRLoopStatistics to
be able to return meaningful results. In return, we get more
compile-time efficient and precise results from the utility.

Ideally, we should be passing HIRLoopStatistics as an argument to the
utility
to make the dependency explicit but this requires making it available at
the
call sites which is non-trivial work especially because it is being used
in
VPlan codebase. As a hack, I have added HIRLoopStatistics as a
dependency to
some passes which use domination utilities.

CMPLRLLVM-48140