Resolve of merge (conflict) 7dd387d2971d [clang] Add __builtin_isfpclass 'main' -> 'xmain-web' (#14089)
CONFLICT (content): Merge conflict in
clang/include/clang/Basic/Builtins.def
  CONFLICT (content): Merge conflict in clang/lib/Sema/SemaChecking.cpp

  commit 7dd387d2971d7759cadfffeb2082439f6c7ddd49
  Author: Serge Pavlov <sepavloff@gmail.com>
  Date:   Sun Jun 18 22:53:32 2023 +0700

      [clang] Add __builtin_isfpclass

A new builtin function __builtin_isfpclass is added. It is called as:

          __builtin_isfpclass(<floating point value>, <test>)

and returns an integer value, which is non-zero if the floating point
argument falls into one of the classes specified by the second argument,
and zero otherwise. The set of classes is an integer value, where each
value class is represented by a bit. There are ten data classes, as
      defined by the IEEE-754 standard, they are represented by bits:

          0x0001 (__FPCLASS_SNAN)         - Signaling NaN
          0x0002 (__FPCLASS_QNAN)         - Quiet NaN
          0x0004 (__FPCLASS_NEGINF)       - Negative infinity
          0x0008 (__FPCLASS_NEGNORMAL)    - Negative normal
          0x0010 (__FPCLASS_NEGSUBNORMAL) - Negative subnormal
          0x0020 (__FPCLASS_NEGZERO)      - Negative zero
          0x0040 (__FPCLASS_POSZERO)      - Positive zero
          0x0080 (__FPCLASS_POSSUBNORMAL) - Positive subnormal
          0x0100 (__FPCLASS_POSNORMAL)    - Positive normal
          0x0200 (__FPCLASS_POSINF)       - Positive infinity

They have corresponding builtin macros to facilitate using the builtin
      function:

if (__builtin_isfpclass(x, __FPCLASS_NEGZERO | __FPCLASS_POSZERO) {
            // x is any zero.
          }

The data class encoding is identical to that used in llvm.is.fpclass
      function.

      Differential Revision: https://reviews.llvm.org/D152351

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: sys_ce_bb <sys_ce_bb@intel.com>