Resolve of merge (conflict) 6f2e92c10ceb Re-land [LLD] Allow usage of LLD as a library 'main' -> 'xmain-web' (#14094)
CONFLICT (content): Merge conflict in lld/tools/lld/lld.cpp

  commit 6f2e92c10cebca51f9bc98b99fa3b5583ff396c4
  Author: Alexandre Ganea <alex_toresh@yahoo.fr>
  Date:   Mon Jun 19 07:32:34 2023 -0400

      Re-land [LLD] Allow usage of LLD as a library

      This reverts commit aa495214b39d475bab24b468de7a7c676ce9e366.

As discussed in https://github.com/llvm/llvm-project/issues/53475 this
patch
allows for using LLD-as-a-lib. It also lets clients link only the
drivers that
      they want (see unit tests).

This also adds the unit test infra as in the other LLVM projects. Among
the
      test coverage, I've added the original issue from @krzysz00, see:
https://github.com/ROCmSoftwarePlatform/D108850-lld-bug-reproduction

Important note: this doesn't allow (yet) linking in parallel. This will
come a
      bit later hopefully, in subsequent patches, for COFF at least.

      Differential revision: https://reviews.llvm.org/D119049

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Ji, Jinsong <jinsong.ji@intel.com>