Resolve of merge (conflict) d46d9689f7c3 [BBUtils] Don't add 'then' block to a loop if it's termina.. 'main' -> 'xmain-web' (#14099)
CONFLICT (content): Merge conflict in
llvm/lib/Transforms/Utils/BasicBlockUtils.cpp

  commit d46d9689f7c3fe16010cd07761858a6392239b3f
  Author: Dmitry Makogon <d.makogon@g.nsu.ru>
  Date:   Thu Jun 8 18:56:18 2023 +0700

[BBUtils] Don't add 'then' block to a loop if it's terminated with
unreachable

      SplitBlockAndInsertIfThen utility creates two new blocks,
they're called ThenBlock and Tail (true and false destinations of a
conditional
branch correspondingly). The function has a bool parameter Unreachable,
      and if it's set, then ThenBlock is terminated with an unreachable.
At the end of the function the new blocks are added to the loop of the
split
block. However, in case ThenBlock is terminated with an unreachable,
      it cannot belong to any loop.

      Differential Revision: https://reviews.llvm.org/D152434

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Jinsong Ji <jinsong.ji@intel.com>