Resolve of merge (build) 8784b6a8540f [Clang] Allow bitcode linking when the input is LLVM-IR 'main' -> 'xmain-web' (#14101)
commit 8784b6a8540f4a333690e7233587859f1d82f620
  Author: Joseph Huber <jhuber6@vols.utk.edu>
  Date:   Wed Jun 7 13:11:19 2023 -0500

      [Clang] Allow bitcode linking when the input is LLVM-IR

Clang provides the `-mlink-bitcode-file` and `-mlink-builtin-bitcode`
      options to insert LLVM-IR into the current TU. These are usefuly
primarily for including LLVM-IR files that require special handling to
be correct and cannot be linked normally, such as GPU vendor libraries
like `libdevice.10.bc`. Currently these options can only be used if the
source input goes through the AST consumer path. This patch makes the
changes necessary to also support this when the input is LLVM-IR. This
      will allow the following operation:

      ```
      clang in.bc -Xclang -mlink-builtin-bitcode -Xclang libdevice.10.bc
      ```

      Reviewed By: yaxunl

      Differential Revision: https://reviews.llvm.org/D152391

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Jinsong Ji <jinsong.ji@intel.com>