[NFCI] Move substitution strings to Diag.cpp to support future translation (#13971)
TCEs have indicated a likely requirement for xmain diagnostics and
remarks to be translated into Japanese. All remarks that may need 
to be translated should be kept in a single module to facilitate this. 
This patch creates an auxiliary string table for strings that don't 
require a remark ID.  The table is initially populated with strings from
the vectorizer that provide reasons for not vectorizing a loop.