[HIRTempCleanup] Bugfix in region liveout load temp's invalidation logic (#14126)
In case we have a candidate load temp which is liveout of region and we
found
its single use in a copy inst, we try to make the the copy's lval temp
as
liveout instead by propagating the load to its rval. We have logic to
invalidate
the candidate if the copy lval is redefined in the region. For example-

t1 = A[i1]; // t1 is liveout of region
t2 = t1; // t2 is the substitute temp where we can propagate t1's rval

t2 =        // t2 is redefined so we cannot make it liveout

The invalidation logic doesn't work when copy's lval temp does not have
a
valid blob index. This happens when the temp is liveout of the region
and
has no uses inside the region. We now detect this case on encountering
the copy
and invalidate the load temp candidate.

CMPLRLLVM-48888