code extractor fix for simd (#14071)
CodeExtractor creates in-memory variables for values that are live-out
from region and passes the memory addresses to the call. In case of
extracting from a simd region, private clauses are created for such
variables. Vectorizer then privatizes the variables creating vectors in
memory and, scalarizing the call, vectorizer needs to pass addresses of
individual items of those vectors. Getting the address of a vector
element is impossible if the items are of the non C-standard size (i.e.
when they are i1, i2, i11 etc).
This PR fixes the problem creating those variables with the nearest
bigger standard size (e.g. i1, i2, i4 as i8, i11 as i16 etc). The
corresponding zext/trunc instructions are emitted before/after
store/load of them. The transformation is restricted for simd regions
only by a special flag passed in the CodeExtractor constructor.

JIRA: CMPLRLLVM-48595