[NFCI] Convert OptReportStatsTracker to use OptRemark for variadic arguments (#14119)
We've previously converted other remarks in the vectorizer to use OptRemark
directly to support any number of arguments.  The OptReportStatsTracker
still uses remarks with one fixed argument, and an empty string is used
for that argument when it's unused, with no support for multiple arguments.
Convert the RemarkRecord to store an OptRemark instead for full
generality.

I opened CMPLRLLVM-48431 to track this when I wanted to use multiple
arguments for OptRemarkID::VectorizerReductionInfo. Change this remark to
use multiple arguments to demonstrate the utility.

Reference: CMPLRLLVM-48431