CMPLRLLVM-48525: Disable temp substitution during HIR dead store elimination when the parent chain can't be proven (#14157)
This patch disables the temp substituion during HIR dead store
elimination
when the parent chain can't be analyzed. The issue is that it clears
the data structure holding the loads to be substituted, and it may
produce a pass when analyzing the refs in this structure. The
consequences
are that not all loads are substituted correctly, and may produce
incorrect code.