Resolve of merge (conflict) 200cc952a28a [LTO][GlobalDCE] Use pass parameter instead of module flag.. 'main' -> 'xmain-web' (#14201)
CONFLICT (content): Merge conflict in llvm/lib/Passes/PassRegistry.def

  commit 200cc952a28a73687ba24d5334415df6332f2d5b
  Author: Teresa Johnson <tejohnson@google.com>
  Date:   Fri Jun 23 11:01:22 2023 -0700

[LTO][GlobalDCE] Use pass parameter instead of module flag for LTO phase

D63932 added a module flag to indicate that we are executing the regular
LTO post merge pipeline, so that GlobalDCE could perform more aggressive
optimization for Dead Virtual Function Elimination. This caused issues
trying to reuse bitcode that had already been through the LTO pipeline
      (see context in D139816).

      Instead support this by passing down a parameter flag to the
GlobalDCEPass constructor, which is the more usual way for indicating
      this information.

Most test changes are to remove incidental uses of this flag. Of the 2
real uses, llvm/test/LTO/ARM/lto-linking-metadata.ll is now obsolete and
removed in this patch, and the virtual-functions-visibility-post-lto.ll
      test is updated to use the regular LTO default pipeline where this
      parameter is set to true.

      Differential Revision: https://reviews.llvm.org/D153655

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Chenyang Liu <chenyang.liu@intel.com>