[RYL] add experimental post-pass to unfuse FMAs in cyclic dependences (#13905)
As discussed in email, this is an initial attempt to consider cyclic
dependencies in single-block loops for fma latency optimizations. For
our royal KRs we'd like to add this to evaluate the performance gains,
where we see a substantial improvement for cpu2017 imagick_s.

The performance improvement is the result of changing a cycle of two
dependent fmas with a cycle of two adds, fed by multiplies. This reduces
the cyclic critical path in half, which is particularly important due to
royal's clustering, but may potentially improve performance for any
uarch with fast adds.

This is fit into global-fma to reuse some of the code, and so that
global-fma doesn't revert any optimized chains. Later, we'll look into
incorporating this into new machine combiner patterns.