Use the MAX_TC_EST of the original loop to set the MAX_TC_EST of the stripmined chunk if it is lower than stripmine size (#14210)
MAX_TC_EST of stripmined loops was not getting updated correctly when it
is less than stripmine size. Ideally speaking, this should be a
performance issue but in this case it resulted in an out-of-range index
created by vectorizer when it unrolls the loop. Locality analysis
(loopopt in general) is not capable of handling out-of-range indices so
we hit an assert. I don't have a proper solution for the out-of-range
index issue but it is a good thing to keep MAX_TC_EST updated with
transformations which is what this PR is doing. Updating MAX_TC_EST
fixes the assertion because vectorizer makes sure VF * UF does not
exceed MAX_TC_EST.

More context is in the JR comments: CMPLRLLVM-45451