IPPredOpt: Change most probable targets to get perf gain for 523 (#14228)
(CMPLRLLVM-26682): Actually

_ZN11xercesc_2_721NameDatatypeValidator7compareEPKtS2_PNS_13MemoryManagerE
is the most probable target at runtime. Change the heuristics to pick
this target. But there are other possible targets that are exactly same
definition so those targets can’t be distinguished. So, decided to pick
all three of them and then generate runtime checks for all of them.