[NFC] CMPLRLLVM-48215: Add collector for SIMD directives in opt-predicate (#14212)
This patch adds a method for collecting and classifying the SIMD
directives
for a loop. The directives will be identified as PreAndPostLoop (in loop
pre-header and post-exit), Region (at region level), NotSupported
(SIMD directive is not supported for transformation). The change set
is NFC since it doesn't changes the behavior of the program. The test
case is an extra test case that we missed during handling SIMD.

JIRA:  CMPLRLLVM-48215