[VPlan][DivergenceAnalysis][Predicator] Fix for CMPLRLLVM-46200 (#14147)
Background:
An assert was triggered during the DA shape calculation for merge-phis
in the predicator. The problem was happening during the blend processing
of merge-phis, where one merge-phi was and operand for another
merge-phi.
The predicator was using a DenseMap to store these phis and later
calling
updateDivergence() to update their shapes. It's possible that under
certain
memory layouts one phi can be processed before the other and vice versa.
Since recomputeShapes() in DA was previously asserting that all operands
must be defined, when the merge-phi containing the other merge-phi
operand
was updated first, it would lead to an assert because the other phi's
shape
had not yet been updated.

This fix does the following:
1) Relaxes the assert for Phis/Branches in DA during shape recomputation
   because it's possible that we may need to push the users of those
instructions to the worklist so that their shapes may be resolved later.
This maintains consistency with what is done during the initial DA shape
   computation.
2) Predicator is updated to use a MapVector, which operates similary to
   DenseMap but is ordered and that keys are stored twice and removal is
   more expensive. However, the number of entries for this map should be
small anyway because it's based on the number of blocks computed from
the
DominanceFrontier calculation and the predicator never removes these
phis
from the map. The ordering prevents the flaky behavior seen in this JIRA
and the accompanying test will always assert without the DA fix provided
   in this PR.