[VPlan][EarlyExitVec] Introduce standalone transform to handle early-exit loops. (#14175)
This PR introduces transformEarlyExitLoop - a new standalone
VPlan-to-VPlan transform that handles early-exit loops. It is expected
to massage the VPlan CFG and explicitly emit mechanisms to track loop
body execution mask and lane that takes the early-exit. The PR
implements the first step that is performed in this transform. This is
the third patch in the series for the early-exit loop vectorization
feature.

JIRA: CMPLRLLVM-48621
Change-Id: I1264d0e2c318dff730ca6309ed0a97d7cf2db1b5