Revert local changes to ShrinkWrap due to D152213 (#14300)
CMPLRLLVM-49129:

With D152213, both cases can be analyzed. There is no stack access, and
no need to ShrinkWrap. However, since D152213's improvement is external
to useOrDefCSROrFI(), the result of it's analysis can't be reused for
the recently-enabled split region optimization in D42600. It still
thinks bb.2 needs to save and restore, and somehow splits it up.
Looks like xmain-web's result is superior because our analysis is
correct in all cases, but D152213 can use some improvement.