[VPlan] Preliminary support for DA checks in VPlan verifier (#14229)
Related JIRA: https://jira.devtools.intel.com/browse/CMPLRLLVM-45767

Adds checks for divergence analysis in the VPlan verifier. There are two
separate checks, one which verifies that all instructions have a shape
set (i.e. are not Undef), and another which checks that the shape set is
correct. The first check can be run at any point, but for now is
disabled before codegen due to issues with not all insts generated after
the predicator having shapes set. For the second check, the instruction
shapes are recalculated and checked after HCFG construction, after which
it is no longer possible to reliably recompute the shapes due to CFG
changes.

Issue for missing DA shapes:
https://jira.devtools.intel.com/browse/CMPLRLLVM-48955