[VPlan][CloneUtils] Fix crash when cloning plan for masked peel. (#14305)
Clone every block in the VPlan instead of using
VPCloneUtils::cloneBlocksRange(&front(), &back()).

The problem with the previous implementation is that cloneBlocksRange()
uses an SESE iterator when cloning blocks. It is guaranteed by
construction that the front() block is also the entry block, but it is
not guaranteed that the back() block is the exit block. In this case, we
will miss some blocks when cloning, causing a crash when iterating over
each block to remap operands. This was seen when cloning a normalized
VPlan to create a masked VPlan for the vector peel.

This fixes a crash seen when adding enabling vectorized peel in
the following benchmarks: specACCEL2023simdref/550,551,553,555.

JR: CMPLRLLVM-45617.