Resolve of merge (conflict) 691dc2d10d4b [Libomptarget] Begin implementing support for RPC services 'main' -> 'xmain-web' (#14344)
CONFLICT (content): Merge conflict in
openmp/libomptarget/test/CMakeLists.txt

  commit 691dc2d10d4bc5a8e6cea266f201693aee46b40d
  Author: Joseph Huber <jhuber6@vols.utk.edu>
  Date:   Fri Jun 30 22:08:42 2023 -0500

      [Libomptarget] Begin implementing support for RPC services

      This patch adds the intial support for running an RPC server in
libomptarget to handle host services. We interface with the library
provided by the `libc` project to stand up a basic server. We introduce
      a new type that is controlled by the plugin and has each device
intialize its interface. We then run a basic server to check the RPC
      buffer.

This patch does not fully implement the interface. In the future each
plugin will want to define special handlers via the interface to support
things like malloc or H2D copies coming from RPC. We will also want to
allow the plugin to specify t he number of ports. This is currently
      capped in the implementation but will be adjusted soon.

Right now running the server is handled by whatever thread ends up doing
the waiting. This is probably not a completely sound solution but I am
not overly familiar with the behaviour of OpenMP tasks and what would be
required here. This works okay with synchrnous regions, and somewhat
fine with `nowait` regions, but I've observed some weird behavior when
      one of those regions calls `exit`.

      Reviewed By: jdoerfert

      Differential Revision: https://reviews.llvm.org/D154312

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: sys_ce_bb <sys_ce_bb@intel.com>
Co-authored-by: Wenwan Xing <wenwan.xing@intel.com>