Complex math with -fhonor-nans and -ffreestanding options. (#14299)
__(div|mul)(s|d)c3 symbols are generated for both Linux and Windows when
the options -fhonor-nans and -fhonor-nan-compares are used. During
oneMKL testing with non-intel compiler, these symbols are resolved on
Linux by the system libs. But on Windows these symbols cannot be
resolved without the Intel compiler. These instrinsics will ultimately
be expanded to library calls.
The -freestanding option is used to prevent library calls. Therefore,
the intrinsics shouldn’t be generated in the first place.
This patch prevents the generation of these intrinsics on Windows, and
when the -ffreestanding option is used.

Fix for CMPLRLLVM-48554 and CMPLRLLVM-49109.