CMPLRLLVM-48215: Enable unswitching for loops that are inside SIMD regions (#14285)
This patchs enables loop unsiwtching when the loopnest is inside a SIMD
region. A SIMD region is an HLRegion where the first instruction is
the begin SIMD directive, and the last instruction is SIMD end. The
memrefs need to be invariant to the region. This will allow to hoist
the whole condition outside of the SIMD clauses.

Gain: lammps/lj +25%