[VPlan][NFC] Pre-commit test changes for peel loop opt-report stats fix. (#14369)
Opt-report stats are incorrectly reported for vectorized peel: we report
load/stores that are aligned in the main loop as aligned in the
vectorized peel loop as well. This patch pre-commits a test change with
a FIXME indicating this. A follow-up patch will commit the functional
change that fixes this behavior.

JR: CMPLRLLVM-49363.