[NFC] Add static interface for calling the VPlan verifier ad-hoc (#14282)
Related JIRA: https://jira.devtools.intel.com/browse/CMPLRLLVM-49105

Adds a static interface to call into the verifier without constructing a
VPlanVerifier instance in advance. The intent of this interface is to be
used by developers when debugging issues in the compiler, as a way of
ensuring that the VPlan is valid at a desired point.