[VPlan][HIR] Fix setting of loop live outs when reduction final related instructions are hoisted up
We hit HIR verification fails when vector loop is unrolled and reduction partial
sums are used. The issue shows up specifically when we hoist out reduction final
operations outside outer loops when we have a perfect loop nest. The fix marks
all blob symbases in the reduction final operand as live out to account for folding.