[VPlan][SLP] Support for SLP graph build on top of VPlan IR. (#14335)
The change implements graph building for similar VPInstruction and SLP
pattern detection for graphs that have stores as seed.

The result of detection manifests as the cost reduction for scalar Plan,
which eventually gives priority to scalar VPlan over non scalar VPlans
during VF selection.

The existing code, which penalizes VF > 1 Plans if SLP pattern is
detected by other means is not disabled in this change set. This will be
done as a separate change set once the new code covers every case the
existing code triggers for.