[OpenMP][NFC] Fix uninitialized variable reported by sanitizer (#14530)
Though VarSize is definitely uninitialized and passed to
registerTargetGlobalVariable, due to conditions on the other parameters
it doesn't appear to be used there. This seems to be why it is hard
to reproduce. Added zero-initializiation for the variable.

This piece of code has been heavily modified from the upstream
making hard to follow. We should consider trying to clean this up in
a separate change.

Fixes: CMPLRLLVM-49490