[TLI][VPlan] Represent required CPU features for vector library functions in SVML lookup table. (#13914)
VPlan generated SVML calls are explicitly legalized by the specialized
MapIntrinToIml pass. However this does not apply to Fortran-specific
vector library functions like the SIMD random number generators. We
currently emit illegal calls for the target which results in stability
issues. This PR updates TLI to represent required CPU features as an
additional attribute in VectorDesc and uses TTI to check the
availability of the feature when determining if a function is
vectorizable. We also extend VPlan's call pumping feature to non-SVML
calls.

JIRA:
[CMPLRLLVM-48566](https://jira.devtools.intel.com/browse/CMPLRLLVM-48566)