[WinSEH] Disable CFG optimizations when SEH enabled (#14581)
CMPLRLLVM-48712

There might be a better solution for SEH CFG problem and should be fixed
in community as well. So far, I put a patch D154294 to disable the pass
as the beginning. Since community wants for a better solution, I'll just
disable them in xmain to solve the problem first.