CMPLRLLVM-49638: Inlining report fixes (#14574)
The following issues need to be fixed in the inlining report:

(1) Cloning should be handled by an interface around CloneFunction() to
catch more cases.
(2) The metadata inlining report needs to tolerate cases when there is
no function metadata for a function, because the function was created
from scratch. In particular, the temps allocated for compacting for any
function should only be accessed when the metadata has been allocated
for that function.
(3) Compacting info must be copied from original to clone when a
function is cloned.

Changes for (1) induced a difference in the intel partial inlining tests
because the cloning was not previously instrumented. I have updated
those LIT tests.
Changes for (2) are tested by InlineReport165.ll, which includes VPO
Paropt Lowering. Eventually, we should instrument the VPO Paropt
Lowering for the inlining report so that the outlined functions are
recognized explicitly.
Changes for (3) are tested by InlineReport166.ll