Fix compile time issue in OpenMPOptCGSCCPass (#14594)
JIRA:
[CMPLRLLVM-49179](https://jira.devtools.intel.com/browse/CMPLRLLVM-49179)
The problem occurred since in OpenMPOptCGSCCPass a complicity of the
code was estimated by the number of instructions in containing Module.
Since getNumberOfInstructions for a module traverses all the contained
BBs it takes quite a time for heavy modules. For OpenMPOptCGSCCPass code
complexity can be estimated just by summing up the number of
instructions a processed SCC contains and this change does that.
Obtained compile time result after this fix does not differ with the
case when complexity calculations are bypassed.