[VPlan] Add ErrorHandler in VPlan passes in the new pipeline. (#14491)
The FatalErrorHandler parameter was dropped during switching to the new
pass manager. Restoring it with some additions. We need to indicate
not only fatal errors but also simple bailouts. The reason is described
in the JIRA. In short, in the new pipeline any footprints of
the [non]vectorization may be lost before OCL passes reach the kernel
code. So those passes can't detect whether the kernel was vectorized.
The suggested solution is to mark the kernel in a special way when we
were unable to vectorize it. We don't want to know about how that
marking is done so the best solution is to invoke a handler.

https://jira.devtools.intel.com/browse/CMPLRLLVM-43692